requirements.txt
aiogram>=3.5
aiosqlite>=0.20
python-dotenv>=1.0
openpyxl>=3.1
